Updated the main file so that it takes command line arguments for CF template (json), existing input threagile yaml (can be a stub) and output file path for output threagile yaml.
The tool will now take an existing yaml as input and output a yaml that
has the existing yaml contents plus the mapped concepts from
CloudFormation.

Also updated the readme.

Still need to add the ability to check if the translated outputs already
exist in the input file. Currently it will duplicate them (not ideal).

Also need to add later the ability to report to the user which
CloudFormation resources, etc. have been mapped and which have not.

And obviously need to add more mappings over time.